+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpInterfaceMethod.cs
@@ -58,7 +58,15 @@ public class CSharpInterfaceMethod : CSharpMember<CSharpInterfaceMethod>, IHasCS
         configure?.Invoke(param);
         return this;
     }
-    
+
+    public CSharpInterfaceMethod InsertParameter(int index, string type, string name, Action<CSharpParameter> configure = null)
+    {
+        var param = new CSharpParameter(type, name);
+        Parameters.Insert(index, param);
+        configure?.Invoke(param);
+        return this;
+    }
+
     public CSharpInterfaceMethod AddGenericParameter(string typeName)
     {
         var param = new CSharpGenericParameter(typeName);